Fix ADAPT_*_C / ADAPT_*_C_BASE macros responsible.
These macros are used to handle generation of boilerplate code or each
member to add to the adapted sequence. And in the commit
abff92ab657333252c9a7a9e00274de99d294d5c we changed the signature of these
macros to handle generation of proxied object field type deduction, this
change simply fix the related macros.