Swap include guards to `#ifndef`
Compilers implement a multiple-include optimization where they avoid
preprocessing an included file if it is "properly guarded". Compilers
differ as to what constitutes a "properly guarded" file. In particular,
Microsoft Visual Studio does not recognize `#if !defined(GUARD)` for
the optimization. This does not affect the correctness of using Fusion
on Microsoft Visual Studio, but it will increase the time taken to build
any project on this compiler.

This commit changes all include guards to use `#ifndef`, but leaves all
other `#if !defined` preprocessor directives that are not related to
guarding a file.